80 - Moves local storage utils to tableConfigPersistence file
