+++ libs/orchestrator-ui-components/src/lib/components/Table/index.ts
@@ -6,6 +6,7 @@ export * from './DataGridTable';
 export * from './dataGridColumns';
 
 export * from './constants';
+export * from './tableConfigPersistence';
 export * from './tableUtils';
 
 export * from './TableSettingsModal';
